fix: new request form as a dialog
