+++ src/app/src/router/index.ts
@@ -1,7 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import LoginPage from '../LoginPage/LoginPage.vue'
 import HomePage from '../HomePage/HomePage.vue'
-import NewRequestPage from '../NewRequestPage/NewRequestPage.vue'
 import RequestsPage from '../RequestsPage/RequestsPage.vue'
 import ThreadPage from '../ThreadPage/ThreadPage.vue'
 import DatasetsPage from '../DatasetsPage/DatasetsPage.vue'
@@ -20,11 +19,6 @@ const router = createRouter({
       path: '/messages',
       name: 'messages',
       children: [
-        {
-          path: 'new-request',
-          name: 'new request',
-          component: NewRequestPage
-        },
         {
           path: 'requests',
           name: 'my requests',
